requirements.txt
Flask
Pillow
inky[rpi]